+++ _content/tour/algorithms-bits-seven.article
@@ -1,5 +1,5 @@
 Bit operations
-This section provides examples that perform bit manipulation.
+This section provides examples that perform bit operations.
 
 * Is Even Or Odd
