+++ _content/tour/algorithms-numbers.article
@@ -1,5 +1,5 @@
 Number operations
-This section provides examples that perform number manipulation.
+This section provides examples that perform number operations.
 
 * Palindrome
