+++ _content/tour/algorithms-searches.article
@@ -1,20 +1,32 @@
-Searches
-TODO - Searches Description
+Search operations
+This section provides examples that perform search operations.
 
-* Searches
+* Binary Search
 
-** Binary Search
+The sample program implements a function that performs an interative
+binary search against set of integers.
 
-.play algorithms/searches/binarysearch/binarysearch.go
+- See more at [[https://en.wikipedia.org/wiki/Binary_search_algorithm][https://en.wikipedia.org/wiki/Binary_search_algorithm]]
 
-** Interpolation Search
+*Diagram*
+
+    Binary search compares the target value to the middle element of the
+    array. If they are not equal, the half in which the target cannot lie
+    is eliminated and the search continues on the remaining half, again
+    taking the middle element to compare to the target value, and repeating
+    this until the target value is found. If the search ends with the
+    remaining half being empty, the target is not in the array
+
+.play algorithms/searches/binary_iterative.go
+
+* Interpolation Search
 
 .play algorithms/searches/interpolationsearch/interpolationsearch.go
 
-** Jump Search
+* Jump Search
 
 .play algorithms/searches/jumpsearch/jumpsearch.go
 
-** Linear Search
+* Linear Search
 
 .play algorithms/searches/linearsearch/linearsearch.go
