+++ _content/tour/algorithms-slices.article
@@ -1,5 +1,5 @@
 Slice operations
-This section provides examples that perform slice manipulation.
+This section provides examples that perform slice operations.
 
 * Minimum Number
