+++ _content/tour/algorithms-sorting.article
@@ -1,5 +1,5 @@
 Sort operations
-This section provides examples that perform sorting manipulation.
+This section provides examples that perform sorting operations.
 
 * Bubble Sort
