+++ _content/tour/algorithms-strings.article
@@ -1,5 +1,5 @@
 String operations
-This section provides examples that perform string manipulation.
+This section provides examples that perform string operations.
 
 * Permutation
