+++ _content/tour/static/js/values.js
@@ -91,11 +91,11 @@ value('tableOfContents', [{
         'algorithms-slices',
         'algorithms-sorting',
         'algorithms-data',
-        'algorithms-searches',
-        'algorithms-fun',
-        'algorithms-simulations',
     ]
 }]).
+// 'algorithms-searches',
+// 'algorithms-fun',
+// 'algorithms-simulations',
 // translation
 value('translation', {
     'off': 'off',
